+++ src/logger/src/entry/message.rs
