+++ src/logger/src/sink/consumer.rs
