+++ src/logger/src/sink/consumer/class.rs
