+++ src/logger/src/sink/formatter/class.rs
